remove favourites, styles

+++ app/src/controllers/inicioController.js
@@ -8,18 +8,18 @@ spaApp
 
     this.listFavorities = [];
 
+    this.existC = false;
+
     this.search = () =>{
-        console.log("SEARch",this.word);
         inicioService.getCharacters(this.word).then( (result) =>{
-          console.log("TERMINO",result);
           this.listCharacters = result.data.results;
         });
     }
 
     this.getComic = (comic) =>{
       inicioService.getComic(comic).then( (result)=>{
-        console.log("RESU",result.data.results);
         this.modalComic = result.data.results[0];
+        this.existComic(this.modalComic.id);
       });
     }
 
@@ -34,10 +34,44 @@ spaApp
 
     this.setFavourities = (item)=>{
       this.getFavourities();
+      this.existComic(item.id);
 
       let items = this.listFavorities;
-      items.push(item);
-      localStorage.setItem("favourities",JSON.stringify(items));
+
+      if(!this.existC){
+        items.push(item);
+        localStorage.setItem("favourities",JSON.stringify(items));
+        this.existC = true;
+      }
+      
+    }
+
+    this.existComic = (item) =>{
+      let found =  this.listFavorities.find((element)=>{
+        return (element.id==item);
+      });
+      
+      if(typeof found !== "undefined"){
+        this.existC = true;
+      }else{
+        this.existC = false;
+      }
+
+    }
+
+    this.removeFavourite = (item) =>{
+      this.getFavourities();
+
+      let items = this.listFavorities;
+
+      let new_items = items.filter((element)=>{
+        return element.id != item;
+      });
+
+      this.listFavorities = new_items;
+
+      localStorage.setItem("favourities",JSON.stringify(new_items));
+
     }
